Made payments.

+++ src/main/java/com/chapterone/chapteroneapp/models/ChargeRequest.java
@@ -4,12 +4,12 @@ package com.chapterone.chapteroneapp.models;
 
 public class ChargeRequest {
     private String token;
-    private int amount;
+    private double amount;
 
     public ChargeRequest() {
     }
 
-    public ChargeRequest(String token, int amount) {
+    public ChargeRequest(String token, double amount) {
         this.token = token;
         this.amount = amount;
     }
@@ -22,11 +22,11 @@ public class ChargeRequest {
         this.token = token;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
